break out list of platforms into its own file/export
