+++ front-end/src/Platforms.js
@@ -1,0 +1,12 @@
+const platforms = [
+  { value: "", label: "All Platforms" },
+  { value: "Netflix", label: "Netflix" },
+  { value: "Amazon Prime", label: "Amazon Prime" },
+  { value: "Hulu", label: "Hulu" },
+  { value: "HBO", label: "HBO" },
+  { value: "Disney Plus", label: "Disney Plus" },
+  { value: "Crunchyroll", label: "Crunchyroll" },
+  { value: "Other", label: "Other" },
+]
+
+export default platforms;
